Switch over to V2 index management details
