+++ x-pack/platform/plugins/shared/index_management/public/application/sections/home/index_list/create_index/create_index_button.tsx
@@ -12,9 +12,7 @@ import { EuiButton } from '@elastic/eui';
 
 import useObservable from 'react-use/lib/useObservable';
 import { CreateIndexModal } from './create_index_modal';
-import { CreateIndexModalV2 } from './create_index_modal_v2';
 import { useAppContext } from '../../../../app_context';
-import { useIsPlatformIndexManagementV2Enabled } from '../../../../hooks/use_is_platform_index_management_v2_enabled';
 
 export interface CreateIndexButtonProps {
   loadIndices: () => void;
@@ -26,11 +24,8 @@ export const CreateIndexButton = ({ loadIndices, share, dataTestSubj }: CreateIn
   const {
     core: { chrome },
   } = useAppContext();
-  const isPlatformIndexManagementV2Enabled = useIsPlatformIndexManagementV2Enabled();
   const [createIndexModalOpen, setCreateIndexModalOpen] = useState<boolean>(false);
   const createIndexUrl = share?.url.locators.get('SEARCH_CREATE_INDEX')?.useUrl({});
-
-  const IndexModal = isPlatformIndexManagementV2Enabled ? CreateIndexModalV2 : CreateIndexModal;
 
   const activeSolutionId = useObservable(chrome.getActiveSolutionNavId$());
 
@@ -55,7 +50,10 @@ export const CreateIndexButton = ({ loadIndices, share, dataTestSubj }: CreateIn
         />
       </EuiButton>
       {createIndexModalOpen && (
-        <IndexModal closeModal={() => setCreateIndexModalOpen(false)} loadIndices={loadIndices} />
+        <CreateIndexModal
+          closeModal={() => setCreateIndexModalOpen(false)}
+          loadIndices={loadIndices}
+        />
       )}
     </>
   );
