+++ x-pack/platform/plugins/shared/index_management/public/application/sections/home/index_list/create_index/create_index_modal.tsx
@@ -8,6 +8,7 @@
 import React, { Fragment, useCallback, useState } from 'react';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
+import { css } from '@emotion/react';
 import {
   EuiButton,
   EuiButtonEmpty,
@@ -21,10 +22,15 @@ import {
   EuiModalHeaderTitle,
   EuiModalBody,
   EuiModalFooter,
-  EuiScreenReaderOnly,
   EuiSpacer,
   EuiText,
   useGeneratedHtmlId,
+  EuiFlexGroup,
+  EuiFlexItem,
+  useEuiTheme,
+  EuiHorizontalRule,
+  EuiCodeBlock,
+  type UseEuiTheme,
 } from '@elastic/eui';
 
 import { LOOKUP_INDEX_MODE, STANDARD_INDEX_MODE } from '../../../../../../common/constants';
@@ -32,12 +38,18 @@ import { indexModeDescriptions, indexModeLabels } from '../../../../lib/index_mo
 import { createIndex } from '../../../../services';
 import { notificationService } from '../../../../services/notification';
 
-import { isValidIndexName } from './utils';
+import { generateRandomIndexName, isValidIndexName } from './utils';
 
 const INVALID_INDEX_NAME_ERROR = i18n.translate(
   'xpack.idxMgmt.createIndex.modal.invalidName.error',
   { defaultMessage: 'Index name is not valid' }
 );
+
+const modalHeaderStyles = ({ euiTheme }: UseEuiTheme) => css`
+  flex-direction: column;
+  align-items: flex-start;
+  gap: ${euiTheme.size.s};
+`;
 
 export interface CreateIndexModalProps {
   closeModal: () => void;
@@ -46,14 +58,24 @@ export interface CreateIndexModalProps {
 
 export const CreateIndexModal = ({ closeModal, loadIndices }: CreateIndexModalProps) => {
   const modalTitleId = useGeneratedHtmlId();
+  const { euiTheme } = useEuiTheme();
 
-  const indexModeFormRowId = useGeneratedHtmlId({ prefix: 'indexMode' });
+  const [indexName, setIndexName] = useState<string>(generateRandomIndexName);
 
-  const [indexName, setIndexName] = useState<string>('');
   const [indexMode, setIndexMode] = useState<string>(STANDARD_INDEX_MODE);
   const [indexNameError, setIndexNameError] = useState<string | undefined>();
   const [isSaving, setIsSaving] = useState<boolean>(false);
   const [createError, setCreateError] = useState<string | undefined>();
+  const [showApi, setShowApi] = useState(false);
+
+  const apiCode = `PUT ${indexName}
+{
+  "settings": {
+    "index":{
+      "mode":"${indexMode}"
+    }
+  }
+}`;
 
   const putCreateIndex = useCallback(async () => {
     setIsSaving(true);
@@ -98,15 +120,23 @@ export const CreateIndexModal = ({ closeModal, loadIndices }: CreateIndexModalPr
       aria-labelledby={modalTitleId}
       onClose={closeModal}
       initialFocus="[name=indexName]"
-      css={{ width: 450 }}
+      css={css`
+        width: calc(${euiTheme.size.xxxl} * 16);
+      `}
     >
-      <EuiModalHeader>
+      <EuiModalHeader css={modalHeaderStyles}>
         <EuiModalHeaderTitle id={modalTitleId}>
           <FormattedMessage
-            id="xpack.idxMgmt.createIndex.modal.title"
-            defaultMessage="Create index"
+            id="xpack.idxMgmt.createIndex.modal.header.title"
+            defaultMessage="Create your index"
           />
         </EuiModalHeaderTitle>
+        <EuiText color="subdued" size="s">
+          <FormattedMessage
+            id="xpack.idxMgmt.createIndex.modal.header.description"
+            defaultMessage="An index stores and defines the schema of your data."
+          />
+        </EuiText>
       </EuiModalHeader>
       <EuiModalBody>
         {createError && (
@@ -131,108 +161,139 @@ export const CreateIndexModal = ({ closeModal, loadIndices }: CreateIndexModalPr
           </>
         )}
         <EuiForm id="createIndexModalForm" component="form">
-          <EuiFormRow
-            fullWidth
-            label={i18n.translate('xpack.idxMgmt.createIndex.modal.indexName.label', {
-              defaultMessage: 'Index name',
-            })}
-            isDisabled={isSaving}
-            isInvalid={indexNameError !== undefined}
-            error={indexNameError}
-          >
-            <EuiFieldText
-              isInvalid={indexNameError !== undefined}
-              fullWidth
-              name="indexName"
-              value={indexName}
-              onChange={(e) => onNameChange(e.target.value)}
-              data-test-subj="createIndexNameFieldText"
-            />
-          </EuiFormRow>
-          <EuiFormRow
-            id={indexModeFormRowId}
-            fullWidth
-            label={
-              <>
-                {i18n.translate('xpack.idxMgmt.createIndex.modal.indexMode.label', {
+          <EuiFlexGroup>
+            <EuiFlexItem>
+              <EuiFormRow
+                fullWidth
+                label={i18n.translate('xpack.idxMgmt.createIndex.modal.indexName.label', {
+                  defaultMessage: 'Index name',
+                })}
+                isDisabled={isSaving}
+                isInvalid={indexNameError !== undefined}
+                error={indexNameError}
+              >
+                <EuiFieldText
+                  isInvalid={indexNameError !== undefined}
+                  fullWidth
+                  name="indexName"
+                  value={indexName}
+                  onChange={(e) => onNameChange(e.target.value)}
+                  data-test-subj="createIndexNameFieldText"
+                />
+              </EuiFormRow>
+            </EuiFlexItem>
+            <EuiFlexItem grow={false}>
+              <EuiFormRow
+                fullWidth
+                label={i18n.translate('xpack.idxMgmt.createIndex.modal.indexMode.label', {
                   defaultMessage: 'Index mode',
                 })}
-                <EuiScreenReaderOnly>
-                  <span>{`, ${
-                    indexModeLabels[indexMode as keyof typeof indexModeLabels] ?? indexMode
-                  }`}</span>
-                </EuiScreenReaderOnly>
-              </>
-            }
-            isDisabled={isSaving}
-          >
-            <EuiSuperSelect
-              fullWidth
-              hasDividers
-              aria-labelledby={`${indexModeFormRowId}-label`}
-              name="indexMode"
-              valueOfSelected={indexMode}
-              onChange={(mode) => setIndexMode(mode)}
-              data-test-subj="indexModeField"
-              options={[
-                {
-                  value: STANDARD_INDEX_MODE,
-                  inputDisplay: indexModeLabels[STANDARD_INDEX_MODE],
-                  'data-test-subj': 'indexModeStandardOption',
-                  dropdownDisplay: (
-                    <Fragment>
-                      <strong>{indexModeLabels[STANDARD_INDEX_MODE]}</strong>
-                      <EuiText size="s" color="subdued">
-                        <p>{indexModeDescriptions[STANDARD_INDEX_MODE]}</p>
-                      </EuiText>
-                    </Fragment>
-                  ),
-                },
-                {
-                  value: LOOKUP_INDEX_MODE,
-                  inputDisplay: indexModeLabels[LOOKUP_INDEX_MODE],
-                  'data-test-subj': 'indexModeLookupOption',
-                  dropdownDisplay: (
-                    <Fragment>
-                      <strong>{indexModeLabels[LOOKUP_INDEX_MODE]}</strong>
-                      <EuiText size="s" color="subdued">
-                        <p>{indexModeDescriptions[LOOKUP_INDEX_MODE]}</p>
-                      </EuiText>
-                    </Fragment>
-                  ),
-                },
-              ]}
-            />
-          </EuiFormRow>
+                isDisabled={isSaving}
+                css={css`
+                  width: calc(${euiTheme.size.xxxl} * 4);
+                `}
+              >
+                <EuiSuperSelect
+                  fullWidth
+                  hasDividers
+                  name="indexMode"
+                  valueOfSelected={indexMode}
+                  onChange={(mode) => setIndexMode(mode)}
+                  data-test-subj="indexModeField"
+                  options={[
+                    {
+                      value: STANDARD_INDEX_MODE,
+                      inputDisplay: indexModeLabels[STANDARD_INDEX_MODE],
+                      'data-test-subj': 'indexModeStandardOption',
+                      dropdownDisplay: (
+                        <Fragment>
+                          <strong>{indexModeLabels[STANDARD_INDEX_MODE]}</strong>
+                          <EuiText size="s" color="subdued">
+                            <p>{indexModeDescriptions[STANDARD_INDEX_MODE]}</p>
+                          </EuiText>
+                        </Fragment>
+                      ),
+                    },
+                    {
+                      value: LOOKUP_INDEX_MODE,
+                      inputDisplay: indexModeLabels[LOOKUP_INDEX_MODE],
+                      'data-test-subj': 'indexModeLookupOption',
+                      dropdownDisplay: (
+                        <Fragment>
+                          <strong>{indexModeLabels[LOOKUP_INDEX_MODE]}</strong>
+                          <EuiText size="s" color="subdued">
+                            <p>{indexModeDescriptions[LOOKUP_INDEX_MODE]}</p>
+                          </EuiText>
+                        </Fragment>
+                      ),
+                    },
+                  ]}
+                />
+              </EuiFormRow>
+            </EuiFlexItem>
+          </EuiFlexGroup>
+          <EuiSpacer size="l" />
+          <EuiHorizontalRule margin="none" />
         </EuiForm>
       </EuiModalBody>
       <EuiModalFooter>
-        <EuiButtonEmpty
-          onClick={closeModal}
-          disabled={isSaving}
-          data-test-subj="createIndexCancelButton"
-          data-telemetry-id="idxMgmt-indexList-createIndex-cancelButton"
-        >
-          <FormattedMessage
-            id="xpack.idxMgmt.createIndex.modal.cancelButton"
-            defaultMessage="Cancel"
-          />
-        </EuiButtonEmpty>
-        <EuiButton
-          fill
-          disabled={indexNameError !== undefined}
-          isLoading={isSaving}
-          type="submit"
-          onClick={onSave}
-          form="createIndexModalForm"
-          data-test-subj="createIndexSaveButton"
-          data-telemetry-id="idxMgmt-indexList-createIndex-saveButton"
-        >
-          <FormattedMessage
-            id="xpack.idxMgmt.createIndex.modal.saveButton"
-            defaultMessage="Create"
-          />
-        </EuiButton>
+        <EuiFlexGroup direction="column">
+          <EuiFlexGroup justifyContent="spaceBetween">
+            <EuiButtonEmpty
+              onClick={closeModal}
+              disabled={isSaving}
+              data-test-subj="createIndexCancelButton"
+              data-telemetry-id="idxMgmt-indexList-createIndex-cancelButton"
+            >
+              <FormattedMessage
+                id="xpack.idxMgmt.createIndex.modal.cancelButton"
+                defaultMessage="Cancel"
+              />
+            </EuiButtonEmpty>
+            <EuiFlexGroup justifyContent="flexEnd">
+              <EuiButtonEmpty
+                iconSide="left"
+                iconType="code"
+                disabled={isSaving}
+                onClick={() => setShowApi((prev) => !prev)}
+                data-test-subj="createIndexShowApiButton"
+                data-telemetry-id="idxMgmt-indexList-createIndex-showApiButton"
+              >
+                {showApi ? (
+                  <FormattedMessage
+                    id="xpack.idxMgmt.createIndex.modal.hideApiButton"
+                    defaultMessage="Hide API"
+                  />
+                ) : (
+                  <FormattedMessage
+                    id="xpack.idxMgmt.createIndex.modal.showApiButton"
+                    defaultMessage="Show API"
+                  />
+                )}
+              </EuiButtonEmpty>
+              <EuiButton
+                fill
+                disabled={indexNameError !== undefined}
+                isLoading={isSaving}
+                type="submit"
+                onClick={onSave}
+                form="createIndexModalForm"
+                data-test-subj="createIndexSaveButton"
+                data-telemetry-id="idxMgmt-indexList-createIndex-saveButton"
+              >
+                <FormattedMessage
+                  id="xpack.idxMgmt.createIndex.modal.saveIndexButton"
+                  defaultMessage="Create my index"
+                />
+              </EuiButton>
+            </EuiFlexGroup>
+          </EuiFlexGroup>
+          {showApi && (
+            <EuiCodeBlock language="json" isCopyable>
+              {apiCode}
+            </EuiCodeBlock>
+          )}
+        </EuiFlexGroup>
       </EuiModalFooter>
     </EuiModal>
   );
