+++ x-pack/platform/plugins/shared/index_management/public/application/sections/home/index_list/details_page/details_page.tsx
@@ -18,15 +18,12 @@ import { IndexDetailsSection, Section } from '../../../../../../common/constants
 import type { Index } from '../../../../../../common';
 import type { Error } from '../../../../../shared_imports';
 import { loadIndex } from '../../../../services';
-import { useIsPlatformIndexManagementV2Enabled } from '../../../../hooks/use_is_platform_index_management_v2_enabled';
 import { DetailsPageError } from './details_page_error';
 import { DetailsPageContent } from './details_page_content';
-import { DetailsPageContentV2 } from './details_page_content_v2';
 
 export const DetailsPage: FunctionComponent<
   RouteComponentProps<{ indexName: string; indexDetailsSection: IndexDetailsSection }>
 > = ({ location: { search }, history }) => {
-  const isPlatformIndexManagementV2Enabled = useIsPlatformIndexManagementV2Enabled();
   const queryParams = useMemo(() => new URLSearchParams(search), [search]);
   const indexName = queryParams.get('indexName') ?? '';
   const tab: IndexDetailsTabId = queryParams.get('tab') ?? IndexDetailsSection.Overview;
@@ -102,18 +99,6 @@ export const DetailsPage: FunctionComponent<
       <DetailsPageError
         indexName={indexName}
         resendRequest={fetchIndexDetails}
-        navigateToIndicesList={navigateToIndicesList}
-      />
-    );
-  }
-  if (isPlatformIndexManagementV2Enabled) {
-    return (
-      <DetailsPageContentV2
-        index={index}
-        tab={tab}
-        fetchIndexDetails={fetchIndexDetails}
-        history={history}
-        search={search}
         navigateToIndicesList={navigateToIndicesList}
       />
     );
